rename package on nodelifecycle controller to match the path
